Adding 3 stage submission file for ELM
I have refactored my submission scripts for ELM runs into a single run
that uses SLURM dependencies. This script was aimed to make 3 stage runs
easier to submit but more importantly to reduce bugs when changing
configurations. The top of the file lists some next extensions on
this submission file to make it even more safe as the work continues.